customize admin views, replace FloatField with DecimalField

+++ orders/models.py
@@ -10,7 +10,7 @@ class Pizza(models.Model):
     kind = models.CharField(max_length=10, choices=KIND_CHOICES)
     size = models.CharField(max_length=10, choices=SIZE_CHOICES)
     toppings_number = models.IntegerField()
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=10, decimal_places=2)
 
     def __str__(self):
         return f"{self.title} {self.kind} pizza ({self.size}) with {self.toppings_number} toppings"
@@ -25,7 +25,7 @@ class PizzaTopping(models.Model):
 
 class SubAddition(models.Model):
     title = models.CharField(max_length=100)
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=10, decimal_places=2)
 
     def __str__(self):
         return f"{self.title}"
@@ -36,7 +36,7 @@ class Sub(models.Model):
 
     title = models.CharField(max_length=100)
     size = models.CharField(max_length=10, choices=SIZE_CHOICES)
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=10, decimal_places=2)
 
     def __str__(self):
         return f"{self.title} ({self.size})"
@@ -45,7 +45,7 @@ class Sub(models.Model):
 
 class Pasta(models.Model):
     title = models.CharField(max_length=100)
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=10, decimal_places=2)
 
     def __str__(self):
         return f"{self.title}"
@@ -53,7 +53,7 @@ class Pasta(models.Model):
 
 class Salad(models.Model):
     title = models.CharField(max_length=100)
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=10, decimal_places=2)
 
     def __str__(self):
         return f"{self.title}"
@@ -64,7 +64,7 @@ class DinnerPlatter(models.Model):
 
     title = models.CharField(max_length=100)
     size = models.CharField(max_length=10, choices=SIZE_CHOICES)
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=10, decimal_places=2)
 
     def __str__(self):
         return f"{self.title} ({self.size})"
